Machine Learning week 4 course project
The purpose of this project is to train Machine Learning model that will
be capable of recognizing how well people perform particular physical
activity (weight lifting). The data used is from accelerometers on the
belt, forearm, arm, and dumbell of 6 participants - and the goal is to
classify the observation in one of 5 possible categories (1 correct -
and 4 common mistakes).